fix script again

+++ src/ARM/internal/functions/deploy/New-ARMDeployment.ps1
@@ -298,15 +298,17 @@ function New-ARMDeployment {
 
     #region deployment stage
     try {
-      #region remove resource lock (if any) before any deployment or teardown operation starts
-      Write-PipelineLogger -LogType "info" -Message "New-ARMDeployment.ResourceLock.Remove"
-      try {
-        $deploymentService.RemoveResourceGroupLock(
-          $scopeObject
-        )
-        Write-PipelineLogger -LogType "success" -Message "New-ARMDeployment.ResourceLock.Deleted.Success"
-      } catch {
-        Write-PipelineLogger -LogType "error" -Message "New-ARMDeployment.ResourceLock.Deleted.Failed. Details: $($_.Exception.Message)"
+      if ($Scope -eq "resourcegroup") {
+        #region remove resource lock (if any) before any deployment or teardown operation starts
+        Write-PipelineLogger -LogType "info" -Message "New-ARMDeployment.ResourceLock.Remove"
+        try {
+          $deploymentService.RemoveResourceGroupLock(
+            $scopeObject
+          )
+          Write-PipelineLogger -LogType "success" -Message "New-ARMDeployment.ResourceLock.Deleted.Success"
+        } catch {
+          Write-PipelineLogger -LogType "error" -Message "New-ARMDeployment.ResourceLock.Deleted.Failed. Details: $($_.Exception.Message)"
+        }
       }
       #endregion
 
@@ -371,16 +373,18 @@ function New-ARMDeployment {
           }
           #endregion
 
-          Write-PipelineLogger -LogType "info" -Message "New-ARMDeployment.ResourceLock.Create"
-          #region add resource lock
-          try {
-            $deploymentService.CreateResourceGroupLock(
-              $scopeObject,
-              $ResourceLock
-            )
-            Write-PipelineLogger -LogType "success" -Message "New-ARMDeployment.ResourceLock.Create.Success"
-          } catch {
-            Write-PipelineLogger -LogType "error" -Message "New-ARMDeployment.ResourceLock.Create.Failed. Details: $($_.Exception.Message)"
+          if ($Scope -eq "resourcegroup") {
+            Write-PipelineLogger -LogType "info" -Message "New-ARMDeployment.ResourceLock.Create"
+            #region add resource lock
+            try {
+              $deploymentService.CreateResourceGroupLock(
+                $scopeObject,
+                $ResourceLock
+              )
+              Write-PipelineLogger -LogType "success" -Message "New-ARMDeployment.ResourceLock.Create.Success"
+            } catch {
+              Write-PipelineLogger -LogType "error" -Message "New-ARMDeployment.ResourceLock.Create.Failed. Details: $($_.Exception.Message)"
+            }
           }
           #endregion
